Safari-609.1.11.1/LayoutTests/inspector/debugger/evaluateOnCallFrame-exception-expected.txt

Tests exceptions with Debugger.evaluateOnCallFrame.


== Running test suite: Debugger.evaluateOnCallFrame.Exception
-- Running test case: Debugger.evaluateOnCallFrame.Exception
PASS: Should be a runtime error.

-- Running test case: Complete
